+++ js/game.js
@@ -1,11 +1,12 @@
 var player;
 var needCanvasUpdate = true;
 var gameEnded = false;
+var scrolled = false;
 
 // Don't change this
 const TMT_VERSION = {
-	tmtNum: "2.3.5",
-	tmtName: "Cooler and Newer Edition"
+	tmtNum: "2.π",
+	tmtName: "Incrementally Updated"
 }
 
 
@@ -162,11 +163,9 @@ function shouldNotify(layer){
 				return true
 		}
 	}
-	if (tmp[layer].shouldNotify){
-		return tmp[layer].shouldNotify
-	}
-	else 
-		return false
+	 
+	return tmp[layer].shouldNotify
+	
 }
 
 function canReset(layer)
@@ -354,13 +353,16 @@ function canCompleteChallenge(layer, x)
 function completeChallenge(layer, x) {
 	var x = player[layer].activeChallenge
 	if (!x) return
-	if (! canCompleteChallenge(layer, x)){
+	
+	let completions = canCompleteChallenge(layer, x)
+	if (!completions){
 		 player[layer].activeChallenge = null
 		return
 	}
 	if (player[layer].challenges[x] < tmp[layer].challenges[x].completionLimit) {
 		needCanvasUpdate = true
-		player[layer].challenges[x] += 1
+		player[layer].challenges[x] += completions
+		player[layer].challenges[x] = Math.min(player[layer].challenges[x], tmp[layer].challenges[x].completionLimit)
 		if (layers[layer].challenges[x].onComplete) run(layers[layer].challenges[x].onComplete, layers[layer].challenges[x])
 	}
 	player[layer].activeChallenge = null
@@ -439,6 +441,7 @@ function gameLoop(diff) {
 			let layer = OTHER_LAYERS[row][item]
 			if (tmp[layer].autoPrestige && tmp[layer].canReset) doReset(layer);
 			if (layers[layer].automate) layers[layer].automate();
+			player[layer].best = player[layer].best.max(player[layer].points)
 			if (layers[layer].autoUpgrade) autobuyUpgrades(layer)
 		}
 	}
@@ -479,10 +482,11 @@ var interval = setInterval(function() {
 	if (needCanvasUpdate){ resizeCanvas();
 		needCanvasUpdate = false;
 	}
+	tmp.scrolled = document.getElementById('treeTab').scrollTop > 30
 	updateTemp();
 	gameLoop(diff)
 	fixNaNs()
-	adjustPopupTime(diff) 
+	adjustPopupTime(0.05) 
 	ticking = false
 }, 50)